rotcltd renamed to proper name (facepalm)
